improve auto-healing when the names manager changes
In the past, aliases were added almost daily and auto-healing was
programmed to detect names that were gone. The cons of that approach
was that it did not work for other changes to the name manager, for
example a new special-cased SHA1 needed manual intervention, but it
was a good trade-off since they were rare.

Nowadays mappings are not added so often, and doing the entire repo
does not take a lot of time actually, so the new strategy consists on
recomputing all authors for all commits. This way, no matter how we
modify the names manager, the changes are going to be reflected
automatically.