deep revision of heuristics
This patch is the result of extensive research tweaking heuristics and
comparing before/after recomputing the entire database.

Since 2008 we have looked at [...] at the end of every line in the
commit message and CHANGELOGs (for svn commits), but this rule has
proven to be too weak and prone to false positives.

From now on the application looks for [...] at the end of the subject
of the commits message (think first paragraph), and for lines in the
message body matching exactly that pattern (nothing else allowed
before or after the square brackets).

In addition, if the commit was imported from Subversion, the end of
the commit message body is also checked, and the CHANGELOGs as a last
resort, as always.

With this refinement we have been able to delete a good deal of
previous false positives, and the heuristics are less prone to new
ones.

Commit counters stay generally equal. Some credits have been
corrected, but changes from a user point of view are marginal. Please
if your counter changed and you'd like to know why just drop me a
line.