Revert "Merge pull request #61 from kaspth/permalink-week"
Reason: I merged this patch because the PR looked good modulo details
I wanted to polish locally. But once the route was used locally I
realized we need to think a little bit more about it.

Since we cannot deploy master with this, and there is a new name
alias waiting to be deployed, we are going to revert by now.

This reverts commit e5501dcb0a448443fb547a1e8f25c20516b6c303, reversing
changes made to 36432d75e9916e0f0a194349eae47ebef4724550.