moves a hard-coded rule to a false positive
While having this exception as a hard-coded rule yields the correct credit,
as a rule of thumb that is a last resort, generally for the use case in
which Git does not even have the information about the author.

When the author can be extracted, but the square brackets heuristic gives
a false positive, we discard the false positive and let the rest of the
logic do its thing.